+++ tests/jsthemis/test.js
@@ -22,7 +22,7 @@ describe("jsthemis", function(){
 	decrypter = new addon.SecureMessage(peer_keypair.private(), keypair.public());
 	intruder_decrypter = new addon.SecureMessage(intruder_keypair.private(), keypair.public());
 	message = new Buffer("Test Message");
-	it("encrypt/decrypt", function(){	    
+	it("encrypt/decrypt", function(){
 	    encrypted_message = encrypter.encrypt(message);
 	    assert.equal(message.toString(), decrypter.decrypt(encrypted_message).toString());
 	    assert.throws(function(){intruder_decrypter.decrypt(encrypted_message);}, expect_code(addon.FAIL));
@@ -86,7 +86,7 @@ describe("jsthemis", function(){
 	    server_keypair = new addon.KeyPair();
 	    client_id = new Buffer("client");
 	    client_keypair = new addon.KeyPair();
-	    
+
 	    server_session = new addon.SecureSession(server_id, server_keypair.private(), function(id){
 		if(id.toString()=="server")
 		    return server_keypair.public();
@@ -175,6 +175,42 @@ describe("jsthemis", function(){
 
 describe("jsthemis", function(){
     describe("secure cell", function(){
+        describe("key generation", function(){
+            const defaultLength = 32
+            it("generates new key buffer", function(){
+                var masterKey = new addon.SymmetricKey()
+                assert.equal(masterKey.length, defaultLength)
+            })
+            it("generates new instances", function(){
+                // Check that we don't reuse the same object
+                var key1 = new addon.SymmetricKey()
+                var key2 = new addon.SymmetricKey()
+                assert.notDeepEqual(key1, key2)
+                assert.notEqual(key1, key2)
+                // A copy should have the same content,
+                // but it's a distint object
+                var key3 = new addon.SymmetricKey(key2)
+                assert.deepEqual(key3, key2)
+                assert.notEqual(key3, key2)
+                assert.notDeepEqual(key3, key1)
+                assert.notEqual(key3, key1)
+            })
+            it("is able to restore SymmetricKey from bytes", function(){
+                var bytes = Buffer.from("MDRwUzB0NG1aN2pvTEEwdVljRFJ5", "base64")
+                var masterKey = new addon.SymmetricKey(bytes)
+                assert.equal(masterKey.length, bytes.length)
+            })
+            it("throws on empty buffer", function(){
+                assert.throws(() => new addon.SymmetricKey(Buffer.from("")),
+                    expect_code(addon.INVALID_PARAMETER))
+                assert.throws(() => new addon.SymmetricKey(""),
+                    expect_code(addon.INVALID_PARAMETER))
+                assert.throws(() => new addon.SymmetricKey(null),
+                    expect_code(addon.INVALID_PARAMETER))
+                assert.throws(() => new addon.SymmetricKey(undefined),
+                    expect_code(addon.INVALID_PARAMETER))
+            })
+        })
 	message=new Buffer("This is test message");
 	password=new Buffer("This is test password");
 	context=new Buffer("This is test context");
@@ -217,7 +253,7 @@ describe("jsthemis", function(){
 	    context_imprint_intruder_decrypter = new addon.SecureCellContextImprint(new Buffer("This is test password1"));
 	    assert.throws(function(){new addon.SecureCellContextImprint(empty_message)});
 	    context_imprint_enc_data = context_imprint_encrypter.encrypt(message, context);
-	    assert.equal(message.length, context_imprint_enc_data.length);	    
+	    assert.equal(message.length, context_imprint_enc_data.length);
 	    context_imprint_dec_data = context_imprint_decrypter.decrypt(context_imprint_enc_data, context);
 	    assert.equal(message.toString(), context_imprint_dec_data.toString());
 	    context_imprint_dec_data = context_imprint_intruder_decrypter.decrypt(context_imprint_enc_data, context);